chore: code cleanup and refactoring
